add combined BP visualisation functionality
